Deprecate jira provider in favor of atlassian.jira provider
