+++ .github/ISSUE_TEMPLATE/airflow_providers_bug_report.yml
@@ -40,6 +40,7 @@ body:
         - apache-sqoop
         - arangodb
         - asana
+        - atlassian-jira
         - celery
         - cloudant
         - cncf-kubernetes
@@ -63,7 +64,6 @@ body:
         - influxdb
         - jdbc
         - jenkins
-        - jira
         - microsoft-azure
         - microsoft-mssql
         - microsoft-psrp
