+++ airflow/providers/atlassian/jira/provider.yaml
@@ -16,45 +16,39 @@
 # under the License.
 
 ---
-package-name: apache-airflow-providers-jira
-name: Jira
+package-name: apache-airflow-providers-atlassian-jira
+name: Atlassian Jira
 description: |
     `Atlassian Jira <https://www.atlassian.com/>`__
 
 versions:
-  - 3.0.1
-  - 3.0.0
-  - 2.0.5
-  - 2.0.4
-  - 2.0.3
-  - 2.0.2
-  - 2.0.1
-  - 2.0.0
-  - 1.0.2
-  - 1.0.1
   - 1.0.0
 
 dependencies:
   - apache-airflow>=2.2.0
-  - apache-airflow-providers-atlassian-jira
+  - JIRA>1.0.7
 
 integrations:
-  - integration-name: Atlassian Jira (deprecated)
+  - integration-name: Atlassian Jira
     external-doc-url: https://www.atlassian.com/pl/software/jira
     logo: /integration-logos/jira/Jira.png
     tags: [software]
 
 operators:
-  - integration-name: Atlassian Jira (deprecated)
+  - integration-name: Atlassian Jira
     python-modules:
-      - airflow.providers.jira.operators.jira
+      - airflow.providers.atlassian.jira.operators.jira
 
 sensors:
-  - integration-name: Atlassian Jira (deprecated)
+  - integration-name: Atlassian Jira
     python-modules:
-      - airflow.providers.jira.sensors.jira
+      - airflow.providers.atlassian.jira.sensors.jira
 
 hooks:
-  - integration-name: Atlassian Jira (deprecated)
+  - integration-name: Atlassian Jira
     python-modules:
-      - airflow.providers.jira.hooks.jira
+      - airflow.providers.atlassian.jira.hooks.jira
+
+connection-types:
+  - hook-class-name: airflow.providers.atlassian.jira.hooks.jira.JiraHook
+    connection-type: jira
