+++ airflow/providers/atlassian/jira/sensors/__init__.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,14 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use :mod:`airflow.providers.atlassian.jira.operators.jira`."""
-
-import warnings
-
-from airflow.providers.atlassian.jira.operators.jira import JiraOperator  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.atlassian.jira.operators.jira`.",
-    DeprecationWarning,
-    stacklevel=2,
-)
