+++ airflow/providers/jira/CHANGELOG.rst
@@ -24,11 +24,18 @@
 Changelog
 ---------
 
+3.1.0
+.....
+
+.. warning:: This is the last release for jira provider. The provider is discontinued.
+   Please switch to atlassian.jira provider
+
+* ``Deprecate jira provider in favor of atlassian.jira provider  (#25930)``
+
 3.0.1
 .....
 
 Bug Fixes
-~~~~~~~~~
 
 * ``Fix the error caused by passing unused context in JiraSensor.poke (#23352)``
