+++ docs/apache-airflow-providers-atlassian-jira/commits.rst
@@ -1,0 +1,32 @@
+
+ .. Licensed to the Apache Software Foundation (ASF) under one
+    or more contributor license agreements.  See the NOTICE file
+    distributed with this work for additional information
+    regarding copyright ownership.  The ASF licenses this file
+    to you under the Apache License, Version 2.0 (the
+    "License"); you may not use this file except in compliance
+    with the License.  You may obtain a copy of the License at
+
+ ..   http://www.apache.org/licenses/LICENSE-2.0
+
+ .. Unless required by applicable law or agreed to in writing,
+    software distributed under the License is distributed on an
+    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+    KIND, either express or implied.  See the License for the
+    specific language governing permissions and limitations
+    under the License.
+
+
+Package apache-airflow-providers-atlassian-jira
+-----------------------------------------------
+
+`Atlassian Jira <https://www.atlassian.com/>`__
+
+
+This is detailed commit list of changes for versions provider package: ``atlassian-jira``.
+For high-level changelog, see :doc:`package information including changelog <index>`.
+
+
+
+1.0.0
+.....
