+++ docs/spelling_wordlist.txt
@@ -516,6 +516,7 @@ assigment
 ast
 async
 athena
+atlassian
 attr
 attrs
 auth
